Refactor in test project
