+++ tests/Trakx.Fireblocks.ApiClient.Tests/Integration/Base/FireblocksClientTestsBase.cs
@@ -2,7 +2,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Serilog;
 using Trakx.Common.Testing.Configuration;
 
-namespace Trakx.Fireblocks.ApiClient.Tests.Integration;
+namespace Trakx.Fireblocks.ApiClient.Tests.Integration.Base;
 
 [Collection(nameof(ApiTestCollection))]
 public class FireblocksClientTestsBase
